requirements.txt
gradio
requests
datetime
websockets
websocket
asyncio
numexpr
bs4
pysimplegui
g4f
gpt4free
fireworks-ai
langchain
langchain-community
langchain-fireworks
PyCharacterAI
forefront
anthropic
openai
cohere
google-api-python-client